+++ utilities.h
@@ -4,9 +4,7 @@
 #include<sstream>
 #include <random>
 #include <functional>
-#include "network.h"
 
-class network;
 
 bool are_equal_with_tolerance(double lhs, double rhs);
 
@@ -36,7 +34,5 @@ bool are_same_distribution(std::uniform_real_distribution<double> lhs,
 bool are_same_env_functions(const std::function<double(std::vector<double>)> &lhs,
                             const std::function<double(std::vector<double>)> &rhs, int n_repeats = 1000);
 
-///Checks if a network and a function return the same output
-bool net_behaves_like_the_function(const network &n, const std::function<double(std::vector<double>)> &f, int n_repeats = 1000);
 
 #endif // UTILITIES_H
